chore: Add imagekit
This commit adds the imagekit-javascript dependency to the project. The dependency is updated to version 3.0.1 in both the package-lock.json and package.json files. This dependency will be used to handle image-related functionality in the application.

Note: Please remove any meta information like issue references, tags, or author names from the commit message.